pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "literature-topic-agent"
version = "0.1.0"
description = "Local topic-driven literature collection and ranking system."
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12",
  "httpx>=0.27",
  "jinja2>=3.1",
  "loguru>=0.7",
  "pydantic>=2.7",
  "PyYAML>=6.0",
  "scholarly>=1.7,<2",
  "requests>=2.31",
  "tenacity>=8.2",
  "typer>=0.12",
  "pandas>=2.2",
]

[project.scripts]
literature-topic-agent = "apps.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
include = [
  "apps",
  "apps.*",
  "core",
  "core.*",
  "domain",
  "domain.*",
  "download",
  "download.*",
  "exporters",
  "exporters.*",
  "parse",
  "parse.*",
  "providers",
  "providers.*",
  "storage",
  "storage.*",
  "summarize",
  "summarize.*",
  "topic",
  "topic.*",
  "workflows",
  "workflows.*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
